Move NavControllerTestUtil to root
